refactor: add list participants hook
