+++ frontend/src/hooks/index.js
@@ -3,3 +3,4 @@ export * from "./useToast";
 export * from "./useForm";
 export * from "./useDeleteParticipant";
 export * from "./useSaveParticipant";
+export * from "./useListParticipants";
